Move panel components into a single folder
